more plotting details

+++ DMT/external/latex.py
@@ -200,7 +200,11 @@ def clean_tex_files(directory, file_name, keep=(".tex", ".pdf"), regex=False):
         test_fct = test_str
 
     for file_curr in Path(directory).iterdir():
-        if test_fct(file_name, file_curr) and (file_curr.suffix not in keep):
+        if (
+            file_curr.is_file()
+            and test_fct(file_name, file_curr)
+            and (file_curr.suffix not in keep)
+        ):
             file_curr.unlink()
